+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerAdapterRegistry.java
@@ -1,6 +1,5 @@
 package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
