+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerMappingRegistry.java
@@ -1,7 +1,5 @@
-package camp.nextstep;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.ArrayList;
